pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "notebookx"
version = "0.2.0"
description = "A lightweight python notebook"
authors = [{ name = "Aditya Godse", email = "adimail2404@gmail.com" }]
license = { text = "CC BY 4.0 (Attribution-NonCommercial 4.0 International)" }
readme = "README_PYPI.md"
requires-python = ">=3.9"
dependencies = [
    "tornado",
    "pyzmq",
    "pre-commit",
    "ipykernel",
    "nbformat",
]

[project.urls]
Homepage = "https://github.com/adimail/notebook-x"
Documentation = "https://adimail.github.io/notebook-x"

[project.scripts]
notebookx = "server.run:main"

[tool.setuptools]
packages = ["server", "frontend"]
